Fixup more harmless C++ warnings
In particular, this addresses the following messages:

from `icc -Wextra`:
* warning #68: integer conversion resulted in a change of sign (many instances)
* include/sympack/lapack.hpp: warning #186: pointless comparison of unsigned integer with zero (two instances)
* mpi_interf.hpp: warning #186: pointless comparison of unsigned integer with zero (two instances)

from cce/9.0.1: (which subsequently fails to build for other, more fundamental reasons)
* include/sympack/impl/DistSparseMatrix_impl.hpp:554:13: error: ordered comparison between pointer and zero ('int *' and 'int')

These changes are sufficient to cleanly pass Intel C++ 19.0 -Wall -Wextra.

Note this is still NOT sufficient to cleanly pass Gnu g++'s stronger -Wall
which still generates nearly a thousand warnings (all of which are
-Wsign-compare, -Wunknown-pragma, -Wreorder, -Wswitch or
-Wdelete-non-virtual-dtor)

+++ include/sympack/SuperNode.hpp
@@ -29,7 +29,7 @@ namespace symPACK{
     bool Last;
     Int GIndex;
     size_t Offset;
-    NZBlockDesc():GIndex(-1),Offset(-1),Last(false){};
+    NZBlockDesc():GIndex(-1),Offset((size_t)-1),Last(false){};
     NZBlockDesc(Int aGIndex, size_t aOffset):GIndex(aGIndex),Offset(aOffset),Last(false){};
   };
